requirements.txt
# =============================================================================
# SHK-CMS Backend Requirements
# =============================================================================

# Core Django Framework
Django==4.2.16
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.10  # Updated version with better wheel support

# Configuration Management
python-decouple==3.8

# CORS Headers for Angular Frontend
django-cors-headers==4.3.1

# Image Processing
Pillow==10.4.0

# Authentication & JWT
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.27.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Environment & Deployment
gunicorn==21.2.0
python-dotenv==1.0.0

# Utilities
django-filter==23.4
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# File Management
django-storages==1.14.2
boto3==1.34.0

# Caching & Tasks (optional)
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0

# Monitoring & Logging
sentry-sdk==1.38.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Excel/CSV Export
openpyxl==3.1.2
# pandas==2.2.0  # Commented out - install separately with: pip install --only-binary=all pandas

# Date/Time Utilities
python-dateutil==2.8.2